requirements.txt
# Core AI Framework
crewai==0.86.0
crewai-tools==0.17.0

# LLM and AI
openai==1.54.3
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7

# Web Interface
streamlit==1.39.0

# Document Processing
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
pandas==2.2.3

# Environment Management
python-dotenv==1.0.1

# Additional Tools
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0